data_processing/requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.3.0
antlr4-python3-runtime==4.7.2
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.2.0
Automat==22.10.0
backcall==0.2.0
baseconvert==1.0.0a4
beautifulsoup4==4.12.3
best-download==0.1.2
bitsandbytes==0.38.1
black==23.3.0
blinker==1.8.1
blis==0.7.11
bolt-python==0.7.1
boto==2.49.0
boto3==1.28.11
botocore==1.31.11
bottle==0.12.25
bs4==0.0.2
cachetools==5.3.1
catalogue==2.0.10
category-encoders==2.6.3
certifi==2023.7.22
cffi==1.16.0
chardet==5.1.0
charset-normalizer==3.2.0
clarabel==0.7.1
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
cloudpickle==3.0.0
cmake==3.27.0
colorama==0.4.6
comm==0.1.3
confection==0.1.4
constantly==23.10.4
construct==2.5.3
contourpy==1.2.1
cryptography==42.0.5
cssselect==1.2.0
cvxpy==1.4.3
cycler==0.12.1
cymem==2.0.8
Cython==3.0.0
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DataProperty==1.0.1
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
deepspeed==0.9.2
deprecation==2.1.0
dill==0.3.6
docker-pycreds==0.4.0
dyNET38==2.1
ecos==2.0.13
exceptiongroup==1.2.0
executing==1.2.0
ExifRead==3.0.0
factorization==1.0.0
fastjsonschema==2.19.1
fibonacci==1.0.0
filelock==3.12.2
fiona==1.9.6
fire==0.5.0
Flask==3.0.3
flatbuffers==2.0.7
fonttools==4.51.0
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
geomdl==5.3.1
Geometry==0.0.23
geopandas==0.14.4
gitdb==4.0.10
GitPython==3.1.32
gmpy2==2.1.5
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
gravdataset @ https://aids-download.aoss.cn-sh-01.sensecoreapi-oss.cn/packages/pypi/gravdataset.tar.gz#sha256=d411477403e1b806991c3b75304d8e752b617f780716504967c8fd53265559a2
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
haversine==2.8.1
haystack==0.42
hjson==3.1.0
huggingface-hub==0.16.4
hyperlink==21.0.0
idna==3.4
imageio==2.34.1
imbalanced-learn==0.12.2
imohash==1.0.5
importlib_metadata==7.1.0
incremental==22.10.0
infinity==1.5
iniconfig==2.0.0
intervals==0.9.2
ipykernel==6.25.0
ipython==8.14.0
ipywidgets==8.1.2
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.2.0
jax==0.3.6
jaxlib==0.3.0
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonlines==3.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab_widgets==3.0.10
kaleido==0.2.1
keras==2.13.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
latex2sympy2==1.9.1
lazy_loader==0.4
Levenshtein==0.25.1
libclang==16.0.6
lightgbm==4.3.0
lit==16.0.6
llvmlite==0.42.0
lm-eval @ git+https://github.com/EleutherAI/lm-evaluation-harness.git@1736d78dd9615107e68ec7f74043b02d4ab68d12
loralib==0.1.1
lxml==5.2.1
marisa-trie==1.1.0
Markdown==3.4.4
MarkupSafe==2.1.3
matlab==0.1
matplotlib==3.7.5
matplotlib-inline==0.1.6
mbstrdecoder==1.1.3
mixpanel==4.10.1
mmh3==4.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.10
mypy==1.10.0
mypy-extensions==1.0.0
nagisa==0.2.8
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
nltk==3.8.1
numba==0.59.1
numexpr==2.8.4
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.27.6
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.3
osqp==0.6.5
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.0
parse==1.20.1
parsel==1.9.1
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
pathvalidate==2.5.2
patsy==0.5.6
pefile==2023.2.7
peft @ git+https://github.com/huggingface/peft.git@c42968617b25a98a6a1c206b041846c826eeaf1a
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
Planetarium==0.0.1
platformdirs==3.9.1
plotly==5.22.0
plotly-resampler==0.10.0
pluggy==1.5.0
pmdarima==2.0.4
portalocker==2.7.0
preshed==3.0.9
primefactors==1.0.0
prompt-toolkit==3.0.39
Protego==0.3.1
protobuf==3.20.0
psutil==5.9.5
ptyprocess==0.7.0
PuLP==2.8.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pyact==0.0.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycaret==3.3.2
pycountry==22.3.5
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.12
PyDispatcher==2.0.7
pydot==2.0.0
pyfra==0.3.1
pyglet==2.0.15
Pygments==2.15.1
pyod==1.1.3
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyproj==3.6.1
PySocks==1.7.1
pytablewriter==0.64.2
pytest==8.2.0
python-dateutil==2.8.2
python-ptrace==0.9.9
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
qdldl==0.1.7.post2
queuelib==1.6.2
rapidfuzz==3.9.0
referencing==0.35.1
regex==2023.6.3
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
responses==0.18.0
retrying==1.3.4
rouge-score==0.1.2
rpds-py==0.18.0
rsa==4.9
s3transfer==0.6.1
sacrebleu==1.5.0
safetensors==0.3.1
schemdraw==0.15
scikit-base==0.7.7
scikit-image==0.23.2
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
Scrapy==2.11.1
scs==3.2.4.post1
seaborn==0.13.2
selenium==4.19.0
sentencepiece==0.1.99
sentry-sdk==1.28.1
service-identity==24.1.0
setproctitle==1.3.2
shapely==2.0.4
six==1.16.0
sktime==0.26.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlitedict==2.1.0
srsly==2.4.8
stack-data==0.6.2
statsmodels==0.14.2
supply==0.0.1
sympy==1.13.2
tabledata==1.3.1
tbats==1.1.3
tcolorpy==0.1.3
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
Theano==1.0.5
thinc==8.2.3
threadpoolctl==3.2.0
thunder-python==1.4.2
tifffile==2024.4.24
tk==0.1.0
tldextract==5.1.2
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.0.1
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.3.2
tqdm==4.65.0
tqdm-multiprocess==0.0.11
traitlets==5.9.0
Transcrypt==3.9.1
transformers==4.31.0
trio==0.25.0
trio-websocket==0.11.1
triton==2.0.0
tsdownsample==0.1.3
Twisted==24.3.0
typepy==1.3.1
typer==0.9.4
typing_extensions==4.11.0
tzdata==2023.3
urllib3==1.26.16
varint==1.0.2
w3lib==2.1.2
wandb==0.15.4
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.4
Werkzeug==3.0.2
widgetsnbextension==4.0.10
wrapt==1.15.0
wsproto==1.2.0
wurlitzer==3.1.0
xxhash==3.2.0
yarl==1.9.2
yaspin==3.0.2
yellowbrick==1.5
zipcodes==1.2.0
zipp==3.18.1
zope.interface==6.3
zstandard==0.21.0
